prevent display of permissions in tab when panel open
