+++ src/app/views/query-runner/request/permissions/Permission.tsx
@@ -14,7 +14,6 @@ import { FormattedMessage, injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 
-import { telemetry } from '../../../../../telemetry';
 import { IPermission, IPermissionProps, IPermissionState } from '../../../../../types/permissions';
 import * as permissionActionCreators from '../../../../services/actions/permissions-action-creator';
 import { translateMessage } from '../../../../utils/translate-messages';
@@ -22,7 +21,6 @@ import { classNames } from '../../../classnames';
 import PanelList from './PanelList';
 import { permissionStyles } from './Permission.styles';
 import TabList from './TabList';
-import { setConsentedStatus } from './util';
 
 export class Permission extends Component<IPermissionProps, IPermissionState> {
 
@@ -65,23 +63,6 @@ export class Permission extends Component<IPermissionProps, IPermissionState> {
       || nextProps.consentedScopes !== this.props.consentedScopes
       || nextState.permissions !== this.state.permissions;
     return shouldUpdate;
-  }
-
-  public searchValueChanged = (event: any, value?: string): void => {
-    const { scopes } = this.props;
-    let filteredPermissions = scopes.data;
-    if (value) {
-      const keyword = value.toLowerCase();
-
-      filteredPermissions = scopes.data.filter((permission: IPermission) => {
-        const name = permission.value.toLowerCase();
-        return name.includes(keyword);
-      });
-    }
-
-    this.setState({
-      permissions: filteredPermissions
-    });
   }
 
   public handleConsent = async (permission: IPermission) => {
@@ -227,24 +208,17 @@ export class Permission extends Component<IPermissionProps, IPermissionState> {
         }
       );
     }
-
     return columns;
   }
 
-
-
-
   public render() {
     const classes = classNames(this.props);
-    const { panel, scopes, tokenPresent, consentedScopes } = this.props;
+    const { panel, scopes } = this.props;
     const { pending: loading } = scopes;
-    const { permissions } = this.state;
 
     const {
       intl: { messages },
     }: any = this.props;
-
-    setConsentedStatus(tokenPresent, permissions, consentedScopes);
 
     const selection = new Selection({
       onSelectionChanged: () => {
@@ -268,42 +242,26 @@ export class Permission extends Component<IPermissionProps, IPermissionState> {
         {!loading &&
           <div className={classes.permissions}>
             {!panel && <TabList
-              permissions={permissions}
               columns={this.getColumns()}
               classes={classes}
               renderItemColumn={(item?: any, index?: number, column?: IColumn) =>
                 this.renderItemColumn(item, index, column)}
               renderDetailsHeader={this.renderDetailsHeader}
-
             />}
             {panel &&
               <div data-is-scrollable={true}>
                 <PanelList
                   classes={classes}
-                  permissions={permissions}
                   messages={messages}
                   selection={selection}
                   columns={this.getColumns()}
                   renderItemColumn={(item?: any, index?: number, column?: IColumn) =>
                     this.renderItemColumn(item, index, column)}
-                  searchValueChanged={(event?: React.ChangeEvent<HTMLInputElement>, value?: string) =>
-                    this.searchValueChanged(event, value)}
                   renderDetailsHeader={this.renderDetailsHeader}
                 />
               </div>
             }
           </div>
-        }
-        {permissions && permissions.length === 0 && !loading &&
-          <Label style={{
-            display: 'flex',
-            width: '100%',
-            minHeight: '200px',
-            justifyContent: 'center',
-            alignItems: 'center'
-          }}>
-            <FormattedMessage id='permissions not found' />
-          </Label>
         }
       </div>
     );
